+++ apps/tup-ui/src/pages/Dashboard/Layout.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { ProfileComponent } from '../components/profile';
-import { RequireAuth } from '../components/utils';
+import { ProfileComponent } from '../../components/profile';
+import { RequireAuth } from '../../components/utils';
 
 const Layout: React.FC = () => {
   return (
