+++ apps/tup-ui/src/pages/index.ts
@@ -1,0 +1,4 @@
+export { default as Sidebar } from './Sidebar';
+export { default as Login } from './Login';
+export { default as Dashboard } from './Dashboard';
+export { default as Logout } from './Logout';
